Backend/requirements.txt
Flask
flask-sqlalchemy
psycopg2-binary
python-dotenv
Flask-Cors